requirements.txt
odoo_upgrade @ git+https://github.com/odoo/upgrade-util@master
pdf417gen==0.7.1
